Thu, Jun 30, 2022, 2:21 PM +00:00

+++ pset4/filter-less/helpers.c
@@ -1,4 +1,5 @@
 #include "helpers.h"
+#include <math.h>
 
 // Convert image to grayscale
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
@@ -7,8 +8,8 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            float x = image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j]rbgtBlue/3;
-            double y = round(x);
+            BYTE x = image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j]rbgtBlue/3;
+            int y = round(x);
             image[i][j].rgbtRed = x;
             image[i][j].rgbtGreen = x;
             image[i][j].rgbtBlue = x;
